fix(web): guard tool-outcome parsing against null JSON output
A tool whose completed state.output was the JSON literal `null` (or any
JSON value that parses to a non-object: true, 42, "x", …) crashed the
session page with `Cannot read properties of null (reading 'success')`.

Root cause: parseJsonFailure / formatJsonFailureOutput cast the JSON.parse
result to Record<string, unknown> and read .success on it, but
JSON.parse('null') === null, so null.success threw. The throw bubbled up
through isErrorOutput -> partOutcome -> burstFailureCount's useMemo+filter
(the ActivityBurst failure tally) and tore down the React tree for the
whole session view. Better Stack patterns 487ae241… (12 occurrences) and
ce68779d… (3 occurrences), prod v0.12.4, on /projects/:id/sessions/:id.

Fix: any parsed value that is not a plain object is, by definition, not the
{success:false,error} failure contract -> return null instead of
dereferencing it. Extracted the pure helpers (looksLikeError,
isErrorOutput, partOutcome, batchOutcome, parseJsonFailure,
cleanErrorMessage, formatJsonFailureOutput, ToolOutcome) into a new
tool-outcome.ts with zero React imports so the verdict logic is testable
without loading the 1500-line infrastructure.tsx. infrastructure.tsx
re-exports them so every existing import keeps working.

Regression test pins every non-object JSON shape (null, true, false, 42,
string, array) across parseJsonFailure / formatJsonFailureOutput /
isErrorOutput / partOutcome, and re-checks the failure contract still
parses.
