+++ apps/web/src/features/session/tool/shared/infrastructure.tsx
@@ -515,143 +515,26 @@ export function ToolEmptyState({ message }: { message: string }) {
   );
 }
 
-export function looksLikeError(text: string): boolean {
-  const t = text.trim();
-  if (t.length > 500) return false;
-  if (/^Error:\s/i.test(t)) return true;
-  if (/^([\w._-]+Error|[\w._-]+Exception):\s/i.test(t)) return true;
-  if (/Traceback \(most recent call last\)/i.test(t)) return true;
-  if (/^\s*\[\s*\{[\s\S]*"message"\s*:/.test(t)) return true;
-  return false;
-}
+// ── Tool-outcome + JSON-failure parsing ────────────────────────────────────
+// Pure helpers (no React) live in `./tool-outcome` so they're testable without
+// loading this 1500-line module. Re-exported here to keep every existing
+// `from '@/features/session/tool/shared/infrastructure'` import working —
+// `partOutcome`, `parseJsonFailure`, `isErrorOutput`, `looksLikeError`,
+// `cleanErrorMessage`, `formatJsonFailureOutput`, and the `ToolOutcome` type
+// are all defined there now. See that file for the null-object guard that
+// stops a tool whose output is the literal `null` from crashing turn render
+// (Better Stack patterns `487ae241…` / `ce68779d…`).
+export {
+  looksLikeError,
+  isErrorOutput,
+  partOutcome,
+  parseJsonFailure,
+  cleanErrorMessage,
+  formatJsonFailureOutput,
+  type ToolOutcome,
+} from './tool-outcome';
 
-/**
- * True when a completed tool's output is an error rather than a result —
- * either a plain error/traceback string or the `{success:false,error}` contract.
- * Tools whose own parser returns nothing should fall back to rendering the
- * error (via `ToolOutputFallback`) instead of a blank panel.
- */
-export function isErrorOutput(output: string): boolean {
-  return !!output && (looksLikeError(output) || parseJsonFailure(output) !== null);
-}
-
-export type ToolOutcome = 'ok' | 'partial' | 'failed';
-
-/**
- * Whether a settled tool call succeeded, half-succeeded, or failed.
- *
- * Three different shapes mean "this failed", and only the first of them used to
- * reach the row's icon:
- *   - the call threw            — `state.status === 'error'`
- *   - the call RETURNED the error — completed, output is `Error: …` or the
- *     `{success:false,error}` contract
- *   - the call returned a batch  — completed, `results[]` with some or all
- *     items `success:false` (scrape_webpage over three URLs, one dead host)
- *
- * Every tool renderer hardcodes its own leading icon — `GlobeIcon` for scrape,
- * a favicon for web_fetch, a terminal for bash — and none of them read the
- * output, so shapes two and three kept drawing "web page" over a failure. This
- * is the one predicate they all share; see {@link ToolOutcomeContext}.
- *
- * A running call is `ok`: it has no verdict yet. The shimmer already says
- * "in flight", and painting a warning on unfinished work is a worse lie than
- * the one this fixes.
- */
-export function partOutcome(part: Pick<ToolPart, 'state'>): ToolOutcome {
-  const state = part.state;
-  if (state.status === 'error') return 'failed';
-  if (state.status !== 'completed') return 'ok';
-
-  const output = typeof state.output === 'string' ? state.output.trim() : '';
-  if (!output) return 'ok';
-  if (isErrorOutput(output)) return 'failed';
-  return batchOutcome(output);
-}
-
-/**
- * A batch tool's per-item verdicts collapsed into one outcome.
- *
- * `scrape_webpage` returns `{total, successful, failed, results:[{success}]}`.
- * Two good pages and one dead host is a real and common result that is neither
- * "ok" nor "failed" — it earns amber, not red. Counted from `results[]` rather
- * than the sibling `failed` field because that field is optional, while the
- * array is the thing actually rendered in the card.
- */
-function batchOutcome(output: string): ToolOutcome {
-  if (!output.startsWith('{')) return 'ok';
-
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(output);
-  } catch {
-    return 'ok';
-  }
-
-  const results = (parsed as { results?: unknown } | null)?.results;
-  if (!Array.isArray(results) || results.length === 0) return 'ok';
-
-  const failed = results.filter(
-    (item) =>
-      item !== null &&
-      typeof item === 'object' &&
-      (item as { success?: unknown }).success === false,
-  ).length;
-
-  if (failed === 0) return 'ok';
-  return failed === results.length ? 'failed' : 'partial';
-}
-
-export function parseJsonFailure(output: string): ParsedJsonFailure | null {
-  const trimmed = output.trim();
-  if (!trimmed) return null;
-
-  let parsed: Record<string, unknown>;
-  try {
-    parsed = JSON.parse(trimmed) as Record<string, unknown>;
-  } catch {
-    return null;
-  }
-
-  if (parsed.success !== false || typeof parsed.error !== 'string') return null;
-
-  const result: ParsedJsonFailure = {
-    errorSummary: parsed.error.trim(),
-    hint: typeof parsed.hint === 'string' ? parsed.hint.trim() : undefined,
-  };
-
-  const nestedMatch = result.errorSummary.match(/:\s*(\{[\s\S]*\})\s*$/);
-  if (!nestedMatch) return result;
-
-  try {
-    const nested = JSON.parse(nestedMatch[1]) as Record<string, unknown>;
-    if (typeof nested.message === 'string' && nested.message.trim()) {
-      result.nestedMessage = nested.message.trim();
-    }
-    if (typeof nested.status === 'number') {
-      result.status = nested.status;
-    }
-    if (typeof nested.error === 'boolean') {
-      result.nestedError = nested.error;
-    }
-  } catch {}
-
-  return result;
-}
-
-/**
- * Normalize a backend error string for display: strip the noisy, often
- * duplicated `Error:` prefixes (e.g. "Error: … Error: Error: …") and collapse
- * whitespace, so the panel shows one calm sentence instead of stack-y jargon.
- * Falls back to the trimmed original if cleaning would empty the string.
- */
-export function cleanErrorMessage(raw: string): string {
-  const cleaned = raw
-    .replace(/^(?:\s*Error:\s*)+/i, '')
-    .replace(/(?:\bError:\s*){2,}/gi, '')
-    .replace(/\s+/g, ' ')
-    .trim();
-  return cleaned || raw.trim();
-}
+import { cleanErrorMessage } from './tool-outcome';
 
 export function JsonFailureOutputCard({
   failure,
@@ -694,44 +577,6 @@ export function JsonFailureOutputCard({
       )}
     </div>
   );
-}
-
-export function formatJsonFailureOutput(output: string): string | null {
-  const trimmed = output.trim();
-  if (!trimmed) return null;
-
-  let parsed: Record<string, unknown>;
-  try {
-    parsed = JSON.parse(trimmed) as Record<string, unknown>;
-  } catch {
-    return null;
-  }
-
-  const success = parsed.success;
-  const error = parsed.error;
-  const hint = parsed.hint;
-
-  if (success !== false || typeof error !== 'string') return null;
-
-  const lines: string[] = [];
-  lines.push(error.trim());
-
-  const nestedMatch = error.match(/:\s*(\{[\s\S]*\})\s*$/);
-  if (nestedMatch) {
-    try {
-      const nested = JSON.parse(nestedMatch[1]) as Record<string, unknown>;
-      const nestedMessage = nested.message;
-      if (typeof nestedMessage === 'string' && nestedMessage.trim()) {
-        lines.push(`Details: ${nestedMessage.trim()}`);
-      }
-    } catch {}
-  }
-
-  if (typeof hint === 'string' && hint.trim()) {
-    lines.push(`Hint: ${hint.trim()}`);
-  }
-
-  return lines.join('\n\n');
 }
 
 export function ToolOutputFallback({
